// frontend2/RentingServiceFrontend/src/app/interfaces/createForRentPostDto.ts
export interface createForRentPostDto {
    Title: string;
    Description: string;
    MainCategory: string;
    SleepingPlaceCount: number;
    Price: number;
    PicturesPath: string;
    Lat: string;
    Lng: string;
    Features: string[];
    Categories: string[];
    BuildingNumber: string;
    Street: string;
    District: string;
    City: string;
    Country: string;
}